+++ src/app/admin/components/adminFilters/adminFilters.component.spec.ts
@@ -45,7 +45,7 @@ import { MockComponent, MockProvider } from "ng-mocks";
 
 import { AdminFilters } from "./adminFilters.component";
 import { AuthService } from "@app/services/auth.service";
-import { Loader } from "@app/components/loader/loader.component";
+import { Loader } from "@common/loader/loader.component";
 import { mockAuthedUser } from "@tests/mockData";
 import { ApiClientService } from "@app/services/apiClient.service";
 import { AdminService } from "@app/services/admin.service";